+++ psi4/src/psi4/libpsio/aiohandler.h
@@ -110,24 +110,24 @@ public:
     ///
     /// These functions are not necessary for psio, but for aio they are.
     ///
-    unsigned long read_discont(size_t unit, const char *key, double **matrix,
+    size_t read_discont(size_t unit, const char *key, double **matrix,
       size_t row_length, size_t col_length, size_t col_skip, psio_address start);
     /// Same as read_discont, but for writing
-    unsigned long write_discont(size_t unit, const char *key, double **matrix,
+    size_t write_discont(size_t unit, const char *key, double **matrix,
       size_t row_length, size_t col_length, size_t col_skip, psio_address start);
 
     /// Zero disk
     /// Fills a double precision disk entry with zeros
     /// Total fill size is rows*cols*sizeof(double)
     /// Buffer memory of cols*sizeof(double) is used
-    unsigned long zero_disk(size_t unit, const char* key, size_t rows, size_t cols);
+    size_t zero_disk(size_t unit, const char* key, size_t rows, size_t cols);
 
     /// Write IWL
     /// Write an IWL buffer, thus containing
     /// IWL_INTS_PER_BUF integrals, 4 labels per integral, plus one
     /// integer indicating whether it is the last buffer and one integer
     /// counting the number of integrals in the current buffer
-    unsigned long write_iwl(size_t unit, const char* key, size_t nints,
+    size_t write_iwl(size_t unit, const char* key, size_t nints,
                             int lastbuf, char* labels, char* values, size_t labsize,
                             size_t valsize, size_t* address);
     /// Generic function bound to thread internally
